backend/requirements.txt
langchain
ollama
uvicorn
fastapi
PyPDF2
python-dotenv
python-multipart
sentence-transformers
scikit-learn
numpy